[update] Jenkins用のBuildコマンドと競合してたので修正
