+++ Assets/Scripts/Editor/GitHubActionsBuildCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 291572d3abc0ded489bf782fefa88e1e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
